backend/requirements.txt
flask
flask-cors
python-dotenv
google-cloud-vision
google-cloud-aiplatform